fix(chat): render markdown tables correctly
