+++ .github/workflows/cla.yml
@@ -1,0 +1,49 @@
+name: CLA
+
+on:
+  issue_comment:
+    types: [created]
+  pull_request_target:
+    types: [opened, synchronize, reopened, closed]
+
+permissions:
+  actions: write
+  contents: write
+  pull-requests: write
+  statuses: write
+
+jobs:
+  cla:
+    runs-on: ubuntu-latest
+    if: >
+      (github.event.issue.pull_request || github.event_name == 'pull_request_target') &&
+      github.actor != 'dependabot[bot]'
+    steps:
+      - name: Check or request CLA signature
+        if: >
+          (github.event.comment.body == 'recheck' ||
+           startsWith(github.event.comment.body, 'I have read the CLA Document and I hereby sign the CLA')) ||
+          github.event_name == 'pull_request_target'
+        uses: contributor-assistant/github-action@v2.6.1
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          path-to-document: https://github.com/dennisadriaans/openrun/blob/main/CLA.md
+          branch: main
+          allowlist: dennisadriaans,cursoragent,bot*,*[bot]
+          path-to-signatures: signatures/v1/cla.json
+
+          custom-notsigned-prcomment: >
+            Thanks for the pull request. Before it can be merged we need a
+            Contributor License Agreement on file.
+
+            Open Run is dual-licensed — AGPLv3 here, and commercial terms for
+            organisations that cannot use AGPL. That is only possible if a single
+            party can license the whole codebase, which is what the CLA
+            establishes. You keep every right to your own contribution.
+
+            Please read the [CLA](https://github.com/dennisadriaans/openrun/blob/main/CLA.md),
+            then comment on this PR with exactly:
+          custom-pr-sign-comment: "I have read the CLA Document and I hereby sign the CLA"
+          custom-allsigned-prcomment: "All contributors have signed the CLA. Thanks."
+          lock-pullrequest-aftermerge: false
