+++ AGENTS.md
@@ -70,9 +70,8 @@ HTTP polling is only the **fallback** when a stream is unhealthy. That is why ho
 **`EventSource` is not the judge of liveness — the heartbeat is.** A socket that dies while
 the machine sleeps stays `readyState === OPEN` and never fires `error`, which would pin
 `streamHealthy` to `true` and switch every fallback poll off for good. `lib/liveStream.ts`
-owns both constants: `SERVER_PING_MS` is the heartbeat both SSE factories import (they
-re-export it rather than declaring their own), and a stream silent past `STALE_AFTER_MS` —
-derived from it — is closed and redialled. Do not reintroduce a second copy of the period.
+owns that: both SSE routes ping on `DEFAULT_PING_MS`, and a stream silent past
+`STALE_AFTER_MS` is closed and redialled. Change one of those constants and change the other.
 Neither hook may open an `EventSource` of its own.
 
 ## Hard rules
