--- changelog.d/automations-running-status.md
@@ -1,2 +1,0 @@
-- The Automations list no longer loads every run's turn events just to draw
-  the in-progress dot. It asks for the live task ids and nothing else.
