+++ changelog.d/user-message-markdown.md
@@ -1,0 +1,3 @@
+You no longer see pasted Markdown tables, code fences, and emphasis as raw
+punctuation after submitting a message. Your prompts now render as Markdown in
+the conversation.
