+++ src/components/workspace/WorkspaceBreadcrumb.tsx
@@ -8,8 +8,7 @@
  */
 import { useNavigate } from '@tanstack/react-router'
 import { Check, ChevronDown, ChevronRight, GitBranch, Plus } from 'lucide-react'
-import { useRef, useState, type ReactNode } from 'react'
-import { useClickOutside } from '../../hooks/useClickOutside'
+import { useEffect, useRef, useState, type ReactNode } from 'react'
 import type { WorkspaceWithMeta } from '../../fns'
 import { fetchLatestRunForWorkspace } from '../../lib/queries'
 
@@ -53,8 +52,22 @@ function BranchSwitcher({
   const [visibleCount, setVisibleCount] = useState(BRANCH_PAGE_SIZE)
   const ref = useRef<HTMLDivElement>(null)
   const navigate = useNavigate()
-  const close = () => setOpen(false)
-  useClickOutside(open, close, [ref])
+
+  useEffect(() => {
+    if (!open) return
+    const onPointer = (e: MouseEvent) => {
+      if (!ref.current?.contains(e.target as Node)) setOpen(false)
+    }
+    const onKey = (e: KeyboardEvent) => {
+      if (e.key === 'Escape') setOpen(false)
+    }
+    document.addEventListener('mousedown', onPointer)
+    document.addEventListener('keydown', onKey)
+    return () => {
+      document.removeEventListener('mousedown', onPointer)
+      document.removeEventListener('keydown', onKey)
+    }
+  }, [open])
 
   const ready = workspaces.filter((w) => w.status === 'ready' || w.id === current.id)
   const visibleWorkspaces = ready.slice(0, visibleCount)
