+++ src/lib/liveStream.ts
@@ -14,7 +14,7 @@
  * backoff, because those are exactly the moments a zombie socket surfaces.
  */
 
-/** Server heartbeat period. Both SSE factories import this; do not duplicate. */
+/** Server heartbeat period; keep in sync with `DEFAULT_PING_MS` on both SSE routes. */
 export const SERVER_PING_MS = 15_000
 /** Silence past this means the socket is dead even if the browser disagrees. */
 export const STALE_AFTER_MS = SERVER_PING_MS * 3 - 5_000
