+++ src/server/files.test.ts
@@ -1,22 +1,9 @@
 import assert from 'node:assert/strict'
-import {
-  mkdtempSync,
-  mkdirSync,
-  readFileSync,
-  realpathSync,
-  rmSync,
-  symlinkSync,
-  writeFileSync,
-} from 'node:fs'
+import { mkdtempSync, mkdirSync, readFileSync, rmSync, writeFileSync } from 'node:fs'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { afterEach, describe, it } from 'node:test'
-import {
-  putWorkspaceFile,
-  resolveInsideWorkspace,
-  workspaceRelPath,
-  writeWorkspaceFile,
-} from './files.ts'
+import { putWorkspaceFile, writeWorkspaceFile } from './files.ts'
 
 const dirs: string[] = []
 
@@ -61,39 +48,5 @@ describe('writeWorkspaceFile', () => {
   it('does not create a missing file', () => {
     const root = makeRoot()
     assert.throws(() => writeWorkspaceFile(root, 'missing.ts', 'x\n'), /not found/)
-  })
-})
-
-describe('resolveInsideWorkspace', () => {
-  it('accepts a nested path inside the workspace', () => {
-    const root = makeRoot()
-    mkdirSync(join(root, 'src'))
-    writeFileSync(join(root, 'src', 'a.ts'), 'ok\n')
-    assert.equal(workspaceRelPath(root, 'src/a.ts'), 'src/a.ts')
-    assert.equal(resolveInsideWorkspace(root, 'src/a.ts'), join(realpathSync(root), 'src', 'a.ts'))
-  })
-
-  it('refuses ../ traversal', () => {
-    const root = makeRoot()
-    mkdirSync(join(root, 'inside'))
-    assert.throws(() => resolveInsideWorkspace(root, '../outside.ts'), /escapes/)
-  })
-
-  it('refuses an absolute path outside the workspace', () => {
-    const root = makeRoot()
-    assert.throws(() => resolveInsideWorkspace(root, '/etc/passwd'), /escapes/)
-  })
-
-  it('refuses a NUL in the path', () => {
-    const root = makeRoot()
-    assert.throws(() => resolveInsideWorkspace(root, 'foo\0bar.ts'), /Invalid path/)
-  })
-
-  it('refuses a symlink that points outside the workspace', () => {
-    const root = makeRoot()
-    const outside = makeRoot()
-    writeFileSync(join(outside, 'secret.txt'), 'nope\n')
-    symlinkSync(outside, join(root, 'link'))
-    assert.throws(() => resolveInsideWorkspace(root, 'link/secret.txt'), /escapes/)
   })
 })
